[241224]  전투 코드 정리
각 상대 배열에 0 일경우 다음 상대 서치
상점 코드 정리
승패에 따른 재화 획득
