+++ ConsoleUiSetting.cs
@@ -6,15 +6,13 @@ using System.Threading.Tasks;
 
 namespace FleetComander_Console
 {
-    internal class ConsoleUiSetting
+    class ConsoleUiSetting
     {
         public void ConsoleShowList()
         {
             Console.WriteLine($"플레이어 함대 : {GameInfo.playerFleet[0].Name} / {GameInfo.playerFleet[1].Name} / {GameInfo.playerFleet[2].Name}");
             Console.WriteLine($"플레이어 함대 : {GameInfo.playerFleet[0].UnitCount} / {GameInfo.playerFleet[1].UnitCount} / {GameInfo.playerFleet[2].UnitCount}");
             Console.WriteLine("");
-            Console.WriteLine($"적 함대 : {GameInfo.enemyFleet[0].Name} / {GameInfo.enemyFleet[1].Name} / {GameInfo.enemyFleet[2].Name}");
-            Console.WriteLine($"적 함대 : {GameInfo.enemyFleet[0].UnitCount} / {GameInfo.enemyFleet[1].UnitCount} / {GameInfo.enemyFleet[2].UnitCount}");
             Console.WriteLine("종료 하시려면 엔터를 누르시오.");
 
             ConsoleKeyInfo setkey;
@@ -22,8 +20,29 @@ namespace FleetComander_Console
 
             if (setkey.Key == ConsoleKey.Enter)
             {
-                Console.Clear();
+                ConsoleClear();
+                GameInfo.mainMenu.SelectListMenu();
             }
+        }
+
+        public void ConsoleClear()
+        {
+            Console.SetCursorPosition(0, 0);
+            Console.WriteLine("                                                                                                                                                        " +
+                "                                                                                                                                                                      " +
+                "                                                                                                                                                                      " +
+                "                                                                                                                                                                      " +
+                "                                                                                                                                                                      " +
+                "                                                                                                                                                                      " +
+                "                                                                                                                                                                      " +
+                "                                                                                                                                                                      " +
+                "                                                                                                                                                                      " +
+                "                                                                                                                                                                      " +
+                "                                                                                                                                                                      " +
+                "                                                                                                                                                                      " +
+                "                                                                                                                                                                      ");
+            Console.SetCursorPosition(0, 0);
+
         }
     }
 }
