+++ FleetComander_Console.csproj
@@ -50,7 +50,7 @@
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ShopSetting.cs" />
-    <Compile Include="UnitSetting.cs" />
+    <Compile Include="BattleSetting.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
