+++ Program.cs
@@ -8,6 +8,13 @@ namespace FleetComander_Console
         static void Main(string[] args)
         {
             GameInfo.mainMenu.GameStart();
+
+            //int a = 5;
+            //int b = 2;
+
+            //float c = (a + b) / (float)15;
+
+            //Console.WriteLine(c);
         }
     }
 }
